Change Fish and Estimate Float types to Double
Float does not work with leaf forms

+++ Sources/App/Controllers/EstimateController.swift
@@ -146,14 +146,14 @@ struct EstimateContext: Encodable {
 struct PostEstimateData: Content {
     let eventId: Int
     let name: String
-    let graylingInCm: Float?
-    let graylingInKg: Float?
-    let troutInCm: Float?
-    let troutInKg: Float?
-    let salmonInCm: Float?
-    let salmonInKg: Float?
-    let charInCm: Float?
-    let charInKg: Float?
+    let graylingInCm: Double?
+    let graylingInKg: Double?
+    let troutInCm: Double?
+    let troutInKg: Double?
+    let salmonInCm: Double?
+    let salmonInKg: Double?
+    let charInCm: Double?
+    let charInKg: Double?
 }
 
 extension PostEstimateData: Validatable, Reflectable {
